Removed unused migrations and added compatibility to laravel 5.8

+++ src/database/migrations/2019_03_16_005834_create_user_has_groups_table.php
@@ -14,7 +14,7 @@ class CreateUserHasGroupsTable extends Migration
     public function up()
     {
         Schema::create('user_has_groups', function (Blueprint $table) {
-            $table->integer('user_id', false, true);
+            $table->bigInteger('user_id', false, true);
             $table->integer('group_id', false, true);
             $table->foreign('user_id')
                 ->references('id')
